requirements.txt
streamlit== 1.24.0
pickle-mixin== 1.0.2
pip == 23.3.2
nltk == 3.8.1
rake-nltk == 1.0.6
PyMuPDF==1.23.8
docx==0.2.4
easyocr== 1.7.1
Pillow== 9.5.0
python-docx==1.1.0